# app/static/swagger.yaml
...
swagger: '2.0'
info:
  title: "Your API"
  description: "This is an API"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
  /mechanics/login:
    post:
      tags:
        - Mechanics
      summary: Mechanics login
      description: Mechanics login credentials using email and password
      parameters:
      - in: body
        name: body
        description: Mechanic login credentials
        required: true
        schema:
          type: object
          properties:
            email:
              type: string
            password:
              type: string 
      responses:
        201:
          description: Mechanic login successful
          schema:
            type: object
            properties:
              token:
                type: string
          examples:
            application/json:
              token: "your_token_here"
  /mechanics:
    post:
      tags:
        - Mechanics
      summary: Create a new mechanic
      description: Info to create the mechanic
      parameters:
        - in: body
          name: body
          description: Mechanic information
          required: true
          schema:
            $ref: '#/definitions/MechanicPayload'
      responses:
        201:
          description: mechanic created successfully
          schema:
            $ref: '#/definitions/MechanicResponse'
          examples:
            application/json:
              email: "john@email.com"
              id: 1
              name: "John Doe"
              password: "hashed_password"
              salary: 50000
              tickets: []
        400:
          description: Invalid input
          schema:
            type: object
            properties:
              message:
                type: string
            examples:
              application/json:
                message: "Invalid input"

    get:
      tags:
        - Mechanics
      summary: get all mechanics
      description: retrieve all mechanics
      responses:
        200:
          description: mechanics retrieved successfully
          schema:
            $ref: '#/definitions/MechanicListResponse'

    put:
      tags:
        - Mechanics
      summary: update mechanic
      description: updating mechanic information
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          description: Mechanic information used for update
          required: true
          schema:
            $ref: '#/definitions/MechanicPayload'
      responses:
        201:
          description: mechanic updated successfully
          schema:
            $ref: '#/definitions/MechanicResponse'
          examples:
            application/json:
              email: "john@email.com"
              id: 1
              name: "John Doe"
              salary: 50000
              password: "hashed_password"
              tickets: []
        400:
          description: Invalid input
          schema:
            type: object
            properties:
              message:
                type: string
            examples:
              application/json:
                message: "Invalid input"

    delete:
      tags:
        - Mechanics
      summary: "Enpoint to delete a member"
      description: "Endpoint to delete a member. Requires token authentication, member must be logged in to delete their account"
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Successfully Deleted member"           
          schema:
            $ref: "#/definitions/DeleteMechanicResponse"
          examples:
            application/json:
              message: "successfully deleted mechanic."
  /mechanics/{mechanic_id}:
    get:
      tags:
        - Mechanics
      summary: Get a mechanic by ID
      description: Retrieve a mechanic by their ID
      parameters:
        - name: mechanic_id
          in: path
          required: true
          type: integer
          description: ID of the mechanic to retrieve
      responses:
        200:
          description: Mechanic retrieved successfully
          schema:
            $ref: '#/definitions/MechanicResponse'
...
  /mechanics/search:
    get:
      tags:
        - Mechanics
      summary: Search for mechanics
      description: Search for mechanics by name
      parameters:
        - name: search
          in: query
          required: true
          type: string
          description: Search query (name)
      responses:
        200:
          description: Mechanics retrieved successfully
          schema:
            type: object
            properties:
              mechanics:
                type: array
                items:
                  $ref: '#/definitions/MechanicListResponse'
# ---------------------------------------------------------------------
...
  /serialized-part/paginate?page={page}&per_page={per_page}:
    get:
      tags:
        - Serialized Parts
      summary: Get all serialized parts with pagination
      description: Retrieve all serialized parts with pagination
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
          description: Page number for pagination
        - name: per_page
          in: query
          required: true
          schema:
            type: integer
          description: Number of serialized parts per page
      responses:
        200:
          description: Serialized parts retrieved successfully
          schema:
            type: object
            properties:
              total_parts:
                type: integer
              page:
                type: integer
              per_page:
                type: integer
              parts:
                type: array
                items:
                  $ref: '#/definitions/serializedPartsListResponse'
  /serialized-part:
    post:
      tags:
        - Serialized Parts
      summary: Create a new serialized part
      description: Info to create the serialized part
      parameters:
        - in: body
          name: body
          description: Serialized part information
          required: true
          schema:
            type: object
            properties:
              desc_id:
                type: integer
      responses:
        201:
          description: Serialized part created successfully
          schema:
            type: object
            properties:
              message:
                type: string
                description: Success message
              part:
                type: object
                properties:
                  desc_id:
                    type: integer
                    description: ID of the inventory description
                  id:
                    type: integer
                    description: ID of the serialized part
                  ticket_id:
                    type: integer
                    description: ID of the associated ticket (if any)
          examples:
            application/json:
              message: "Toyota Tires added successfully"
              part:
                desc_id: 2
                id: 5
                ticket_id: null
        400:
          description: Invalid input
          schema:
            type: object
            properties:
              message:
                type: string
            examples:
              application/json:
                message: "Invalid input" 
  /serialized-part/stock/{description_id}:
    get:
      tags:
        - Serialized Parts
      summary: Get all serialized parts by description ID
      description: Retrieve all serialized parts by their description ID
      parameters:
        - name: description_id
          in: path
          required: true
          type: integer
          description: ID of the description to retrieve serialized parts for
      responses:
        200:
          description: Serialized parts retrieved successfully
          schema:
            type: object
            properties:
              Item:
                type: string
                description: name of inventory item
              stock:
                type: integer
                description: number of serialized parts in stock
          examples:
            application/json:
              Item: "part_name"
              stock: 10
  /serialized-part/{serialized_part_id}:
    delete:
      tags:
        - Serialized Parts
      summary: Delete a serialized part by ID
      description: Delete a serialized part by its ID
      parameters:
        - name: serialized_part_id
          in: path
          required: true
          type: integer
          description: ID of the serialized part to delete
      responses:
        200:
          description: Serialized part deleted successfully
          schema:
            $ref: '#/definitions/DeleteInventoryResponse'
          examples:
            application/json:
              message: "successfully deleted serialized part."

# ----------------------------------------------------------------------
...
definitions:
...
  serializedPartsListResponse:
    type: array
    items:
      type: object
      properties:
        id:
          type: integer
        desc_id:
          type: integer
        ticket_id:
          type: integer
  DeleteInventoryResponse:
    type: object
    properties:
      message:
        type: string
...
  DeleteMechanicResponse:
    type: object
    properties:
      message:
        type: string
  MechanicPayload:
    type: object
    properties:
      address:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      salary:
        type: integer
  MechanicResponse:
    type: object
    properties:
      address:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      salary:
        type: integer
      tickets:
        type: array
        items:
          type: integer     
  MechanicListResponse:
    type: array
    items:
      type: object
      properties:
        email:
          type: string
        id:
          type: integer
        name:
          type: string
        password:
          type: string
        salary:
          type: integer
        tickets:
          type: array
          items:
            type: integer